Gestion des droits d'accès sur les pages admin Ok . initFormLogin create methode ValideUser()

+++ application/controllers/AdminController.php
@@ -19,6 +19,7 @@ class AdminController extends CI_Controller {
         parent::__construct();
         $this->load->helper('url');
         $this->load->helper('url_helper');
+        $this->load->library('session');
         $this->load->model('AboutModel', 'aboutM');
         $this->load->model('SkillModel', 'skillsM');
         $this->load->model('CourseModel', 'courM');
@@ -31,43 +32,62 @@ class AdminController extends CI_Controller {
      */
     public function index() {
 
-        $this->load->view('admin/template_admin/head');
-        $this->load->view('admin/template_admin/nav');
-        $this->load->view('admin/home_admin');
-        $this->load->view('admin/template_admin/footer');
+
+
+        if (empty($this->session->userdata('email')) == true) {
+            redirect('login');
+        } else {
+            $this->load->view('admin/template_admin/head');
+            $this->load->view('admin/template_admin/nav');
+            $this->load->view('admin/home_admin');
+            $this->load->view('admin/template_admin/footer');
+        }
+
+//    
     }
 
     public function getAbouts() {
-        $this->load->view('admin/template_admin/head');
-        $this->load->view('admin/template_admin/nav');
-        $this->load->view('admin/about/homeAbout_admin', array(
-            "abouts" => $this->aboutM->getAll(),
-            
-        ));
-        $this->load->view('admin/template_admin/footer');
+
+        if (empty($this->session->userdata('email')) == true) {
+            redirect('login');
+        } else {
+            $this->load->view('admin/template_admin/head');
+            $this->load->view('admin/template_admin/nav');
+            $this->load->view('admin/about/homeAbout_admin', array(
+                "abouts" => $this->aboutM->getAll(),
+            ));
+            $this->load->view('admin/template_admin/footer');
+        }
     }
 
     public function addAbout() {
-        //chargement de la librairie pour la validation du formulaire
-        $this->load->library('form_validation');
-        //chargement du helper form
-        $this->load->helper('form');
         
-        //create Obj about
-        $data = [
-            'text' => $this->input->post('text')
-        ];
-        $this->aboutM->create($data);
-       /// Return on view homeAbout
-        $this->getAbouts();
-     
-        
-    }
-    
-     public function removeAbout($int) {
+        if (empty($this->session->userdata('email')) == true) {
+            redirect('login');
+        } else {
+            //chargement de la librairie pour la validation du formulaire
+            $this->load->library('form_validation');
+            //chargement du helper form
+            $this->load->helper('form');
 
-        $this->aboutM->delete($int);
-        $this->getAbouts();
+            //create Obj about
+            $data = [
+                'text' => $this->input->post('text')
+            ];
+            $this->aboutM->create($data);
+            /// Return on view homeAbout
+            $this->getAbouts();
+        }
+    }
+
+    public function removeAbout($int) {
+        
+        if (empty($this->session->userdata('email')) == true) {
+            redirect('login');
+        } else {
+            $this->aboutM->delete($int);
+            $this->getAbouts();
+        }
     }
 
 }
